+++ UCTB/model_unit/GCLSTM_CELL.py
@@ -158,12 +158,12 @@ class GCLSTMCell(LayerRNNCell):
 
     def KthChebyPloy(self, k, T_k_1=None, T_k_2=None):
         if k == 0:
-            return linalg_ops.eye(self._num_nodes)
+            return linalg_ops.eye(self._num_nodes, dtype=dtypes.float16)
         elif k == 1:
             return self._laplacian_matrix
         elif k > 1:
             return math_ops.matmul(2 * self._laplacian_matrix, T_k_1) - T_k_2
-
+        
     def zero_state(self, batch_size, dtype):
         state_size = self.state_size
         is_eager = context.executing_eagerly()
